# internal/model/subscription.go
package model

import (
	"time"

	"github.com/google/uuid"
)

type Subscription struct {
	ID          int64      `db:"id"`
	ServiceName string     `db:"service_name"`
	Price       int        `db:"price"`
	UserID      uuid.UUID  `db:"user_id"`
	StartDate   time.Time  `db:"start_date"`
	EndDate     *time.Time `db:"end_date"`
	CreatedAt   time.Time  `db:"created_at"`
}

type SubscriptionRequest struct {
	ServiceName string  `json:"service_name" validate:"required"`
	Price       int     `json:"price" validate:"required,gt=0"`
	UserID      string  `json:"user_id" validate:"required,uuid"`
	StartDate   string  `json:"start_date" validate:"required"`
	EndDate     *string `json:"end_date"`
}

type SubscriptionResponse struct {
	ID          int64   `json:"id"`
	ServiceName string  `json:"service_name"`
	Price       int     `json:"price"`
	UserID      string  `json:"user_id"`
	StartDate   string  `json:"start_date"`
	EndDate     *string `json:"end_date,omitempty"`
	CreatedAt   string  `json:"created_at"`
}

type SubscriptionUpdateRequest struct {
	ServiceName *string `json:"service_name"`
	Price       *int    `json:"price"        validate:"omitempty,gt=0"`
	UserID      *string `json:"user_id"      validate:"omitempty,uuid"`
	StartDate   *string `json:"start_date"`
	EndDate     *string `json:"end_date"`
}

type SubscriptionFilter struct {
	UserID      *string
	ServiceName *string
}

type TotalCostFilter struct {
	UserID      *string
	ServiceName *string
	DateFrom    time.Time
	DateTo      time.Time
}